requirements.txt
pandas
plotly-express
pymysql
sqlalchemy
fire
psycopg2
# pytrends
git+https://github.com/jackedison/pytrends.git@add-daily-data-method